tilemaprenderer REGRESSION -> use world coordinates and let the view do the
transformations. no more of that tile offset shit. unfortunately,
this introduces another bug, and in reality this has no real benefit..

i mean, in theory it should be better instead of doing my offset bullshit,
but suddenly now i'm seeing soem really fucked up texture coordinate
bleeding. which shouldn't be happpening at all because i'm clamping
them and using nearest..and it never happened before, and all i changed
was the position of the vertices. puzzling.

i wonder how it would look if the padding was transparent (or i could
discard()).

+++ src/world.h
@@ -151,7 +151,7 @@ private:
 
     void saveMap();
 
-    glm::ivec2 tileOffset(Player* player) const;
+    glm::vec2 tileOffset(Player* player) const;
 
     /**
      * Should be called AFTER the world has been fully processed in raw block form.
